move this back into a service so scott doesnt panic
