--- src/components/wonkLoader.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const WonkLoader: React.FC = () => {
-  return (
-    <div>
-      <h1>Loading...</h1>
-      {/* Add any additional loading indicators or animations here */}
-    </div>
-  );
-};
-
-export default WonkLoader;
